+++ app/helpers/lite/component/component_helper.rb
@@ -4,9 +4,9 @@ module Lite
   module Component
     module ComponentHelper
 
-      def component(name, options = {})
-        name = name.component_path if name.respond_to?(:component_path)
-        "#{name}_component".classify.constantize.render(self, options)
+      def component(name, options = {}, &block)
+        klass = Lite::Component::Base.build(name)
+        klass.render(self, options, &block)
       end
 
     end